change URL for server connection
When running in development, we use :5173 for frontend, and :3000 for
websocket backend.  So we need to tell the client to use :3000 when it
sets up the socket.io client.

When running on the remote server, HTTP and WS traffic goes in through
:80. NGINX handles traffic: pointing HTTP requests to :5173 and WS
requests to :3000.

So to get this to work on the server, we need this commit, which removes
the :3000 from the multiplayer client.

This should probably be handled with an environment variable, but for
now, it'll just be this commit that I constantly rebase onto other work,
lol